Support torch.ScriptObject in torch::jit::as_object
When a torchbind class is returned from an operator, it has the class
`torch.ScriptObject`, yet the `torch.ops` interface checks against
`torch.jit.RecursiveScriptClass` or else falls back to a much slower
path that doesn't return the original c++ object.

On my machine I see a 2 us performance improvement when calling a
`torch.ops` function with a `ScriptObject` argument.

ghstack-source-id: 1085fe3a73313e5a5012e1e1fc837af4cc09cf99
Pull Request resolved: https://github.com/pytorch/pytorch/pull/84398