Contiguify bias in slow_conv_transpose3d kernel
Users never run into this because PyTorch now comes with cudnn by
default and cudnn has a better conv_transpose implementation. However we
seem to test without cudnn in our CI; and also, ROCM goes down this
path.

The .contiguous() call does not regress anything because previously it
was a runtime error. Because this kernel is the "slow conv transpose3d
kernel", we don't care much for its performance.

Test Plan:
- wait for tests

ghstack-source-id: 9882293486d389b2d7444fd26f57170cdab28dac
Pull Request resolved: https://github.com/pytorch/pytorch/pull/84125