[functorch] delete some macros
Pull Request resolved: https://github.com/pytorch/pytorch/pull/84639


This file aliased dispatch keys. The original purpose was so that we
could change the dispatch keys in pytorch core without changing
functorch too much, but there's no need for the layer of indirection
anymore.

Also moved SINGLE_ARG to functorch/csrc/Macros.h, but that might need a
new home later.

Differential Revision: [D39315052](https://our.internmc.facebook.com/intern/diff/D39315052/)
ghstack-source-id: 166807993